cluster/applications/embedding-retrieval/pyproject.toml
[tool.poetry]
name = "embedding-retrieval"
version = "0.1.0"
description = ""
authors = ["Kai Aihara <kaidotio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"

uvicorn = "^0.22.0"
asyncio = "^3.4.3"
aiohttp = "^3.8.4"
fastapi = "^0.100.0"

openai = "^1.33.0"
tiktoken = "^0.7.0"
qdrant-client = "^1.10.1"
fastembed = "^0.3.3"

pydantic = "^1.10.10"

opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-instrumentation-httpx = "^0.45b0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-exporter-prometheus = "^0.45b0"
prometheus-client = "^0.17.0"
python-dotenv = "^1.0.0"
python-json-logger = "^2.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
